requirements.txt
Django==3.0.7
django-crispy-forms==1.9.1
django-heroku==0.3.1
gunicorn==20.0.4
Pillow==7.1.2
requests==2.23.0
